+++ src/test/java/blackjack/domain/machine/HitTest.java
@@ -1,7 +1,6 @@
 package blackjack.domain.machine;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
